monitor_pipeline/requirements.txt
functions-framework==3.*
google-cloud-logging
google-cloud-pubsub